chore(wmg): move Loader to wmgv2
